Modify all cards in stacks
This change iterates through the `_cards` property (the array of cards
in a HuiStackCard), only if it exists, to apply the same mods to each
sub-card in a stack.  Previously, the mods only applied to the first
card in a stack given the `document.querySelector` calls return the
first matching element only.

If the given card isn't a Stack or a CardModder containing a Stack, the
same original behaviour is applied, eg to just the card itself.